Update README.md
Is this the brach of the master? Because I am suppose to edit the readme file of the branch just so I know what a branch is and how commits are done on branches and later on merged with the master files.

This is cool. I do so many edits in my files sometimes I forgot which one is which.